[nim, james] #X set net price to zero if quantity is zero in sales order item and changed order number field to int in sales order

+++ eums/api/import_data/import_sales_orders_endpoint.py
@@ -11,13 +11,14 @@ from eums.vision.vision_facade import SalesOrderFacade
 @csrf_exempt
 @parser_classes((MultiPartParser,))
 def import_sales_orders(request):
-    file_contents = request.FILES['file'].file.getvalue()
-    temp_file = NamedTemporaryFile(mode='w+b', delete=False, suffix='.xlsx')
-    temp_file.write(file_contents)
-    temp_file.close()
+    if 'file' in request.FILES:
+        file_contents = request.FILES['file'].file.getvalue()
+        temp_file = NamedTemporaryFile(mode='w+b', delete=False, suffix='.xlsx')
+        temp_file.write(file_contents)
+        temp_file.close()
 
-    sales_order_facade = SalesOrderFacade(temp_file.name)
-    sales_order_facade.import_orders()
-    os.remove(temp_file.name)
+        sales_order_facade = SalesOrderFacade(temp_file.name)
+        sales_order_facade.import_orders()
+        os.remove(temp_file.name)
 
     return HttpResponse(status=200)
